requirements.txt
streamlit
Pillow
PyPDF2
pandas
openpyxl
xlrd
google-generativeai
altair